Fixes to two specs that were failing intermittantly
For whatever reason, they were consistant for me, so I fixed them.

Specifically:
  spec/requests/admin/page_update_spec.rb
  This had to do with a race between factory definition and use. Instead
  of creating content blocks in the request spec (and the factory
  definition) I let the factory create the blocks and simply refer to
  them.

The other issue had to do with browser sizing. I'm not the only one with
this problem:
http://c.g.renhuali.cn/p/chromedriver/issues/detail?id=1036

Upshot is that width < 348 with certain versions of chrome/chromedriver
results in a reduction to height of 61 == the height of the window
chrome. (address bar etc.)

This was exacerbated by the fact that we had a lingering BrowserSize in
Xing, even though that tool was moved to Waterpig. So fixes kinda worked
and that WP clobbered them.